Started to debug oversize json files

+++ include/json.h
@@ -10,7 +10,7 @@
 #include<errno.h>
 
 struct json_object* getJson(char *str);
-int file_tojson(char* fileName, struct json_object **json);
+long file_tojson(char* fileName, struct json_object **json);
 int getValue(struct  json_object *obj, char *field, char **str);
 struct json_object* getObj(struct  json_object *obj, char *field);
 struct json_object* getObj_rec(struct json_object *obj, char *fields);
